refactor: Create enums/ in data
